\name{mlica2-package}
\alias{mlica2-package}
\alias{mlica2}
\docType{package}
\title{Maximum likelihood implementation of independent component analysis}
\description{An R code implementation of the maximum likelihood (fixed
        point) algorithm of Hyvaerinen, Karhuna, and Oja for independent
        component analysis.}
\details{
\tabular{ll}{
Package: \tab mlica2\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-07-17\cr
License: \tab GPL-2\cr
}
}
\author{Andrew Teschendorff a.teschendorff@ucl.ac.uk (ported to R > 2.10 by Thomas Nelson)

Maintainer: Thomas Nelson
}
\references{
	Hyvaerinen A., Karhunen J., and Oja E.: Independent Component
     Analysis, John Wiley and Sons, New York, (2001).

	Kreil D. and MacKay D. (2003): Reproducibility Assessment of
     Independent Component Analysis of Expression Ratios from DNA
     microarrays, Comparative and Functional Genomics *4* (3),300-317.

	Liebermeister W. (2002): Linear Modes of gene expression determined
     by independent component analysis, Bioinformatics *18*, no.1,
     51-60.

	Chiappetta P., Roubaud MC. and Torresani B.: Blind source separation
     and the analysis of microarray data, J. Comput. Biol. 2004;
     11(6):1090-109.}
	 
\keyword{ package }